Create Assignment-1.cpp
This program calculates the number of boxes required to ship X number of widgets to a customer. 
Have your program prompt the user for the number of widgets ordered. The maximum number of widgets
that can be shipped in a box is 16.  Your program should then calculate the number of full boxes that
will be shipped to the customer and the number of widgets in the last box if the number of widgets
isn't evenly divisible by 16.  You can assume the user will enter in a value between 1 and 15,999.